refactor: #207 keychain 리뷰사항 반영
