+++ ByeBoo-iOS/ByeBoo-iOS/Data/Repository/AuthRepository.swift
@@ -9,13 +9,16 @@ import Foundation
 
 struct DefaultAuthRepository: AuthInterface {
     private let network: NetworkService
+    private let keychainService: KeychainService
     private let userDefaultsService: UserDefaultService
     
     init(
         network: NetworkService,
+        keychainService: KeychainService,
         userDefaultsService: UserDefaultService
     ) {
         self.network = network
+        self.keychainService = keychainService
         self.userDefaultsService = userDefaultsService
     }
     
@@ -23,18 +26,18 @@ struct DefaultAuthRepository: AuthInterface {
     
     func kakaoLogin() async throws{
         let authorization = try await network.request()
-        DefaultKeychainService.save(key: .authorization, token: authorization)
+        keychainService.save(key: .authorization, token: authorization)
     }
     
-    func postLogin(platform: String) async throws {
-        let loginRequestDTO = LoginRequestDTO(platform: platform)
+    func postLogin(platform: LoginPlatform) async throws {
+        let loginRequestDTO = LoginRequestDTO(platform: platform.rawValue)
         let result = try await network.request(
             AuthAPI.login(requestDTO: loginRequestDTO),
             decodingType: PostLoginResponseDTO.self
         )
         _ = userDefaultsService.save(result.isRegistered, key: .isRegistered)
-        DefaultKeychainService.save(key: .accessToken, token: result.accessToken)
-        DefaultKeychainService.save(key: .refreshToken, token: result.refreshToken)
+        keychainService.save(key: .accessToken, token: result.accessToken)
+        keychainService.save(key: .refreshToken, token: result.refreshToken)
     }
     
 }
@@ -44,6 +47,6 @@ struct MockAuthRepository: AuthInterface {
     func kakaoLogin() async throws  {
     }
     
-    func postLogin(platform: String) async throws {
+    func postLogin(platform: LoginPlatform) async throws {
     }
 }
